Função delete completa

+++ cli_students/lib/src/students/subcomands/delete_student_command.dart
@@ -1,8 +1,11 @@
+import 'dart:io';
+
 import 'package:args/command_runner.dart';
 
+import '../../models/students.dart';
 import '../../repositories/product_repositories.dart';
 import '../../repositories/student_repository.dart';
-import 'insert_command.dart';
+// import 'insert_command.dart';
 
 class DeleteById extends Command {
   final StudentRepository studentRepository;
@@ -15,19 +18,35 @@ class DeleteById extends Command {
   String get name => 'delete';
 
   DeleteById(this.studentRepository) {
-    argParser.addOption('file', help: 'Path of the csv file', abbr: 'f');
+    argParser.addOption('delete', help: 'Option', abbr: 'd');
+    argParser.addOption('id', help: 'Insert id', abbr: 'i');
   }
 
   @override
-  void run() {
+  Future<void> run() async {
     print('Aguarde...');
-    final studentId = argResults?['id'];
+    final studentId = int.tryParse(argResults?['id']);
 
     if (studentId == null) {
-      print('Aluno não encontrado.');
+      print('Por favor informe o id do aluno');
       return;
     }
 
-    studentRepository.deleteById(studentId);
+    final student = await studentRepository.findById(studentId);
+
+    print('Voce realmente deseja ecxluir o aluno ${student.name}?  (S) ou (N)');
+    final confirmation = stdin.readLineSync();
+
+    if (confirmation?.toLowerCase() == 's') {
+      await studentRepository.deleteById(studentId);
+      print('===========================');
+      print('Aluno excluído com sucesso');
+      print('===========================');
+    } else {
+      print('===========================');
+      print('Operação Cancelada!!!!');
+      print('===========================');
+      return;
+    }
   }
 }
